+++ src/main/java/at/hannibal2/skyhanni/data/ActionBarData.kt
@@ -3,7 +3,9 @@ package at.hannibal2.skyhanni.data
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.config.commands.CommandCategory
+//#if TODO
 import at.hannibal2.skyhanni.config.commands.CommandRegistrationEvent
+//#endif
 import at.hannibal2.skyhanni.events.ActionBarUpdateEvent
 import at.hannibal2.skyhanni.events.DebugDataCollectEvent
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
@@ -13,6 +15,7 @@ import at.hannibal2.skyhanni.utils.StringUtils.stripHypixelMessage
 import kotlinx.coroutines.launch
 import net.minecraft.util.IChatComponent
 
+// todo 1.21 impl needed
 @SkyHanniModule
 object ActionBarData {
     private var actionBar = ""
@@ -20,6 +23,7 @@ object ActionBarData {
 
     fun getActionBar() = actionBar
 
+    //#if TODO
     @HandleEvent
     fun onCommandRegistration(event: CommandRegistrationEvent) {
         event.register("shtestactionbar") {
@@ -28,6 +32,7 @@ object ActionBarData {
             callback { debugCommand() }
         }
     }
+    //#endif
 
     private fun debugCommand() {
         SkyHanniMod.coroutineScope.launch {
