+++ src/main/java/at/hannibal2/skyhanni/data/ChatManager.kt
@@ -3,11 +3,15 @@ package at.hannibal2.skyhanni.data
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
 import at.hannibal2.skyhanni.config.commands.CommandCategory
+//#if TODO
 import at.hannibal2.skyhanni.config.commands.CommandRegistrationEvent
+//#endif
 import at.hannibal2.skyhanni.events.MessageSendToServerEvent
 import at.hannibal2.skyhanni.events.chat.SkyHanniChatEvent
 import at.hannibal2.skyhanni.events.minecraft.packet.PacketSentEvent
+//#if TODO
 import at.hannibal2.skyhanni.features.chat.ChatHistoryGui
+//#endif
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
 import at.hannibal2.skyhanni.utils.ChatUtils
 import at.hannibal2.skyhanni.utils.DelayedRun
@@ -30,6 +34,7 @@ import kotlin.time.Duration.Companion.seconds
 //$$ import net.minecraft.client.gui.hud.MessageIndicator
 //#endif
 
+// todo 1.21 impl needed
 @SkyHanniModule
 object ChatManager {
 
@@ -175,6 +180,7 @@ object ChatManager {
     }
 
     private fun openChatHistoryGui(args: Array<String>) {
+        //#if TODO
         SkyHanniMod.screenToOpen = if (args.isEmpty()) {
             ChatHistoryGui(getRecentMessageHistory())
         } else {
@@ -186,6 +192,7 @@ object ChatManager {
             }
             ChatHistoryGui(history)
         }
+        //#endif
     }
 
     // TODO: Add another predicate to stop searching after a certain amount of lines have been searched
@@ -265,6 +272,7 @@ object ChatManager {
         }
     }
 
+    //#if TODO
     @HandleEvent
     fun onCommandRegistration(event: CommandRegistrationEvent) {
         event.register("shchathistory") {
@@ -273,4 +281,5 @@ object ChatManager {
             callback { openChatHistoryGui(it) }
         }
     }
+    //#endif
 }
