+++ src/main/java/at/hannibal2/skyhanni/data/IslandType.kt
@@ -5,11 +5,9 @@ import at.hannibal2.skyhanni.api.event.HandleEvent.Companion.HIGHEST
 import at.hannibal2.skyhanni.data.jsonobjects.repo.IslandTypeJson
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
 import at.hannibal2.skyhanni.skyhannimodule.SkyHanniModule
-//#if TODO
 import at.hannibal2.skyhanni.utils.LocationUtils.isInside
 import at.hannibal2.skyhanni.utils.LorenzVec
 import at.hannibal2.skyhanni.utils.SkyBlockUtils
-//#endif
 import net.minecraft.util.AxisAlignedBB
 
 enum class IslandType(private val nameFallback: String) {
@@ -69,9 +67,7 @@ enum class IslandType(private val nameFallback: String) {
 
     val displayName: String get() = islandData?.name ?: nameFallback
 
-    //#if TODO
     fun isInBounds(vec: LorenzVec): Boolean = islandData?.boundingBox?.isInside(vec) ?: true
-    //#endif
 
     @SkyHanniModule
     companion object {
@@ -119,10 +115,8 @@ enum class IslandType(private val nameFallback: String) {
         }
     }
 
-    //#if TODO
     // TODO rename to isInIsland once the funciton in lorenz utils is gone
     fun isCurrent() = SkyBlockUtils.inSkyBlock && SkyBlockUtils.currentIsland == this
-    //#endif
 }
 
 data class IslandData(
