+++ src/main/java/at/hannibal2/skyhanni/data/repo/RepoUtils.kt
@@ -13,7 +13,6 @@ import java.nio.charset.StandardCharsets
 import java.nio.file.Files
 import java.util.zip.ZipInputStream
 
-// todo 1.21 impl needed
 object RepoUtils {
 
     fun recursiveDelete(file: File) {
@@ -108,8 +107,6 @@ object RepoUtils {
 
     @JvmStatic
     fun resetRepoLocation() {
-        //#if TODO
         RepoManager.resetRepositoryLocation(manual = true)
-        //#endif
     }
 }
