+++ src/main/java/at/hannibal2/skyhanni/events/minecraft/packet/PacketSentEvent.kt
@@ -1,14 +1,17 @@
 package at.hannibal2.skyhanni.events.minecraft.packet
 
 import at.hannibal2.skyhanni.api.event.CancellableSkyHanniEvent
-import net.minecraft.client.network.NetHandlerPlayClient
 import net.minecraft.network.Packet
 
-class PacketSentEvent(val network: NetHandlerPlayClient, val packet: Packet<*>) : CancellableSkyHanniEvent() {
+class PacketSentEvent(val packet: Packet<*>) : CancellableSkyHanniEvent() {
     fun findOriginatingModCall(skipSkyhanni: Boolean = false): StackTraceElement? {
         val nonMinecraftOriginatingStack = Thread.currentThread().stackTrace
             // Skip calls before the event is being called
+            //#if MC < 1.21
             .dropWhile { it.className != "net.minecraft.client.network.NetHandlerPlayClient" }
+            //#else
+            //$$ .dropWhile { it.className != "net.minecraft.network.ClientConnection" }
+            //#endif
             // Limit the remaining callstack until only the main entrypoint to hide the relauncher
             .takeWhile { !it.className.endsWith(".Main") }
             // Drop minecraft or skyhanni call frames
