+++ src/main/java/at/hannibal2/skyhanni/utils/json/SkyHanniTypeAdapters.kt
@@ -1,17 +1,17 @@
 package at.hannibal2.skyhanni.utils.json
 
 import at.hannibal2.skyhanni.data.IslandType
-//#if MC < 1.21
+//#if TODO
 import at.hannibal2.skyhanni.data.model.SkyblockStat
 //#endif
 import at.hannibal2.skyhanni.features.fishing.trophy.TrophyRarity
-//#if MC < 1.21
+//#if TODO
 import at.hannibal2.skyhanni.features.garden.CropType
 import at.hannibal2.skyhanni.features.garden.pests.PestType
 //#endif
 import at.hannibal2.skyhanni.utils.LorenzRarity
-//#if MC < 1.21
 import at.hannibal2.skyhanni.utils.LorenzVec
+//#if TODO
 import at.hannibal2.skyhanni.utils.NeuInternalName
 import at.hannibal2.skyhanni.utils.NeuInternalName.Companion.toInternalName
 import at.hannibal2.skyhanni.utils.NeuItems
@@ -19,7 +19,7 @@ import at.hannibal2.skyhanni.utils.NeuItems
 import at.hannibal2.skyhanni.utils.SimpleTimeMark
 import at.hannibal2.skyhanni.utils.SimpleTimeMark.Companion.asTimeMark
 import at.hannibal2.skyhanni.utils.system.ModVersion
-//#if MC < 1.21
+//#if TODO
 import at.hannibal2.skyhanni.utils.tracker.SkyHanniTracker
 //#endif
 import com.google.gson.GsonBuilder
@@ -34,7 +34,7 @@ import kotlin.time.Duration.Companion.milliseconds
 
 // todo 1.21 impl needed
 object SkyHanniTypeAdapters {
-    //#if MC < 1.21
+    //#if TODO
     val NEU_ITEMSTACK: TypeAdapter<ItemStack> = SimpleStringTypeAdapter(NeuItems::saveNBTData, NeuItems::loadNBTData)
     //#endif
 
@@ -43,18 +43,17 @@ object SkyHanniTypeAdapters {
         { java.util.UUID.fromString(this) },
     )
 
-    //#if MC < 1.21
+    //#if TODO
     val INTERNAL_NAME: TypeAdapter<NeuInternalName> = SimpleStringTypeAdapter(
         { this.asString() },
         { this.toInternalName() },
     )
-
+    //#endif
 
     val VEC_STRING: TypeAdapter<LorenzVec> = SimpleStringTypeAdapter(
         LorenzVec::asStoredString,
         LorenzVec::decodeFromString,
     )
-    //#endif
 
     val TROPHY_RARITY: TypeAdapter<TrophyRarity> = SimpleStringTypeAdapter(
         { name },
@@ -81,7 +80,7 @@ object SkyHanniTypeAdapters {
         }
     }
 
-    //#if MC < 1.21
+    //#if TODO
     val CROP_TYPE: TypeAdapter<CropType> = SimpleStringTypeAdapter(
         { name },
         { CropType.getByName(this) },
@@ -100,7 +99,7 @@ object SkyHanniTypeAdapters {
 
     val MOD_VERSION: TypeAdapter<ModVersion> = SimpleStringTypeAdapter(ModVersion::asString, ModVersion::fromString)
 
-    //#if MC < 1.21
+    //#if TODO
     val TRACKER_DISPLAY_MODE = SimpleStringTypeAdapter.forEnum<SkyHanniTracker.DefaultDisplayMode>()
     //#endif
     val ISLAND_TYPE = SimpleStringTypeAdapter.forEnum<IslandType>(IslandType.UNKNOWN)
