+++ src/main/java/at/hannibal2/skyhanni/utils/repopatterns/RepoPatternManager.kt
@@ -2,13 +2,9 @@ package at.hannibal2.skyhanni.utils.repopatterns
 
 import at.hannibal2.skyhanni.SkyHanniMod
 import at.hannibal2.skyhanni.api.event.HandleEvent
-//#if TODO
 import at.hannibal2.skyhanni.config.ConfigManager
-//#endif
 import at.hannibal2.skyhanni.config.features.dev.RepoPatternConfig
-//#if TODO
 import at.hannibal2.skyhanni.data.repo.RepoManager
-//#endif
 import at.hannibal2.skyhanni.events.ConfigLoadEvent
 import at.hannibal2.skyhanni.events.RepositoryReloadEvent
 import at.hannibal2.skyhanni.events.utils.PreInitFinishedEvent
@@ -30,7 +26,6 @@ import net.minecraft.launchwrapper.Launch
 import net.minecraftforge.fml.common.FMLCommonHandler
 //#endif
 
-// todo 1.21 impl needed
 /**
  * Manages [RepoPattern]s.
  */
@@ -84,13 +79,8 @@ object RepoPatternManager {
             }
         }
 
-    //#if TODO
     private val localLoading: Boolean
         get() = config.forceLocal.get() || (!insideTest && PlatformUtils.isDevEnvironment) || RepoManager.usingBackupRepo
-    //#else
-    //$$ private val localLoading: Boolean
-    //$$      get() = config.forceLocal.get() || (!insideTest && PlatformUtils.isDevEnvironment)
-    //#endif
 
     private val logger = LogManager.getLogger("SkyHanni")
 
@@ -267,7 +257,6 @@ object RepoPatternManager {
      * Dump all regexes labeled with the label into the file.
      */
     fun dump(sourceLabel: String, file: File) {
-        //#if TODO
         val data =
             ConfigManager.gson.toJson(
                 RepoPatternDump(
@@ -277,7 +266,6 @@ object RepoPatternManager {
             )
         file.parentFile.mkdirs()
         file.writeText(data)
-        //#endif
     }
 
     @HandleEvent
