+++ versions/1.21.5/buildpaths.txt
@@ -61,7 +61,6 @@ at/hannibal2/skyhanni/skyhannimodule/SkyHanniModule.kt
 at/hannibal2/skyhanni/utils/CircularList.kt
 at/hannibal2/skyhanni/utils/KSerializable.kt
 at/hannibal2/skyhanni/utils/OtherModsSettings.kt
-at/hannibal2/skyhanni/utils/RegexUtils.kt
 at/hannibal2/skyhanni/utils/compat/ArrayNormalization.kt
 at/hannibal2/skyhanni/utils/compat/EffectsCompat.kt
 #at/hannibal2/skyhanni/utils/compat/EnchantmentsCompat.kt
@@ -97,7 +96,6 @@ at/hannibal2/skyhanni/events/utils/PreInitFinishedEvent.kt
 at/hannibal2/skyhanni/skyhannimodule/LoadedModules.kt
 at/hannibal2/skyhanni/utils/compat/MinecraftCompat.kt
 at/hannibal2/skyhanni/utils/RegexUtils.kt
-at/hannibal2/skyhanni/utils/NumberUtil.kt
 at/hannibal2/skyhanni/utils/LorenzColor.kt
 at/hannibal2/skyhanni/utils/system/PlatformUtils.kt
 at/hannibal2/skyhanni/features/misc/visualwords/VisualWord.kt
@@ -593,9 +591,50 @@ at.hannibal2.skyhanni.data.DateFormat.kt
 at.hannibal2.skyhanni.features.inventory.experimentationtable.ExperimentationTableEnums.kt
 at.hannibal2.skyhanni.config.enums.OutsideSBFeature.kt
 at/hannibal2/skyhanni/api/minecraftevents/RenderEvents.kt
-
-
 at.hannibal2.skyhanni.mixins.transformers.MixinMouse.java
-
+at/hannibal2/skyhanni/utils/SkyblockSeason.kt
+at.hannibal2.skyhanni.utils.SkyBlockTime.kt
+at.hannibal2.skyhanni.events.RawScoreboardUpdateEvent.kt
+at.hannibal2.skyhanni.events.ScoreboardUpdateEvent.kt
+at.hannibal2.skyhanni.events.minecraft.ScoreboardTitleUpdateEvent.kt
+at.hannibal2.skyhanni.events.minecraft.packet.PacketReceivedEvent.kt
+at.hannibal2.skyhanni.mixins.transformers.MixinNetworkManager.java
+at.hannibal2.skyhanni.data.ScoreboardData.kt
+at.hannibal2.skyhanni.utils.TimeUtils.kt
+at.hannibal2.skyhanni.utils.RecalculatingValue.kt
+at.hannibal2.skyhanni.events.TabListUpdateEvent.kt
+at.hannibal2.skyhanni.data.model.TabWidget.kt
+at.hannibal2.skyhanni.events.WidgetUpdateEvent.kt
+at.hannibal2.skyhanni.utils.TabListData.kt
+at.hannibal2.skyhanni.mixins.transformers.AccessorGuiPlayerTabOverlay.java
+at.hannibal2.skyhanni.events.TabListLineRenderEvent.kt
+at/hannibal2/skyhanni/mixins/hooks/GuiPlayerTabOverlayHook.kt
+at.hannibal2.skyhanni.test.TestBingo.kt
+at.hannibal2.skyhanni.utils.SkyBlockUtils.kt
+at.hannibal2.skyhanni.events.TablistFooterUpdateEvent.kt
+at.hannibal2.skyhanni.events.hypixel.modapi.HypixelApiJoinEvent.kt
+at.hannibal2.skyhanni.api.hypixelapi.HypixelLocationApi.kt
+at.hannibal2.skyhanni.events.hypixel.HypixelLeaveEvent.kt
+at/hannibal2/skyhanni/events/skyblock/AreaChangeEvents.kt
+at.hannibal2.skyhanni.events.skyblock.SkyBlockLeaveEvent.kt
+at.hannibal2.skyhanni.data.HypixelData.kt
+at.hannibal2.skyhanni.events.hypixel.HypixelJoinEvent.kt
+at.hannibal2.skyhanni.utils.HypixelCommands.kt
+at.hannibal2.skyhanni.data.IslandTypeTag.kt
+at.hannibal2.skyhanni.data.LocationFixData.kt
+at.hannibal2.skyhanni.data.jsonobjects.repo.LocationFixJson.kt
+at.hannibal2.skyhanni.utils.LorenzVec.kt
+at.hannibal2.skyhanni.utils.LocationUtils.kt
+at.hannibal2.skyhanni.data.repo.RepoManager.kt
+at.hannibal2.skyhanni.data.repo.RepoUtils.kt
+at.hannibal2.skyhanni.data.ChatManager.kt
+at.hannibal2.skyhanni.events.minecraft.packet.PacketSentEvent.kt
+at.hannibal2.skyhanni.data.ActionBarData.kt
+at/hannibal2/skyhanni/mixins/hooks/MessageHandlerHook.kt
+at.hannibal2.skyhanni.mixins.transformers.MixinMessageHandler.java
+at.hannibal2.skyhanni.events.ActionBarUpdateEvent.kt
+at/hannibal2/skyhanni/data/ActionBarStatsData.kt
+at.hannibal2.skyhanni.features.chat.SkyblockXPInChat.kt
+at.hannibal2.skyhanni.events.ActionBarValueUpdateEvent.kt
 
 # \.(?!kt|java|\()
