+++ versions/1.21.5/src/main/java/at/hannibal2/skyhanni/mixins/transformers/MixinClientConnection.java
@@ -1,7 +1,9 @@
 package at.hannibal2.skyhanni.mixins.transformers;
 
 import at.hannibal2.skyhanni.events.minecraft.packet.PacketReceivedEvent;
+import at.hannibal2.skyhanni.events.minecraft.packet.PacketSentEvent;
 import net.minecraft.network.ClientConnection;
+import net.minecraft.network.PacketCallbacks;
 import net.minecraft.network.listener.PacketListener;
 import net.minecraft.network.packet.Packet;
 import org.spongepowered.asm.mixin.Mixin;
@@ -15,5 +17,12 @@ public class MixinClientConnection {
     @Inject(method = "handlePacket", at = @At(value = "HEAD"))
     private static void handlePacket$Inject$HEAD(Packet<?> packet, PacketListener listener, CallbackInfo ci) {
         new PacketReceivedEvent(packet).post();
+    }
+
+    @Inject(method = "sendInternal", at = @At(value = "HEAD"), cancellable = true)
+    private void sendPacket(Packet<?> packet, PacketCallbacks callbacks, boolean flush, CallbackInfo ci) {
+        if (new PacketSentEvent(packet).post()) {
+            ci.cancel();
+        }
     }
 }
