+++ versions/1.21.5/src/main/resources/fabric.mod.json
@@ -13,7 +13,15 @@
     ],
     "entrypoints": {
         "main": [
-            "at.hannibal2.skyhanni.SkyHanniModLoader"
+            {
+                "adapter": "kotlin",
+                "value": "at.hannibal2.skyhanni.SkyHanniModLoader"
+            }
         ]
+    },
+    "depends": {
+        "fabric-api": ">=0.119.9+1.21.5",
+        "fabric-language-kotlin": ">=1.13.2+kotlin.2.1.20",
+        "minecraft": ">=1.21.5"
     }
 }
