+++ versions/1.8.9/detekt/baseline.xml
@@ -71,6 +71,17 @@
     <ID>PreprocessingImportOrdering:ChunkedStatsConfig.kt$//#if TODO</ID>
     <ID>PreprocessingImportOrdering:EventsConfig.kt$//#if TODO</ID>
     <ID>PreprocessingImportOrdering:InquisitorSoundConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:HypixelCommands.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:LaneSwitchSoundSettings.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:LowHealthAlertSoundConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:RepositoryConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:SkyHanniMod.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:TitleAndFooterConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:AlignmentConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:BroodmotherSpawnAlertConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:ChunkedStatsConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:EventsConfig.kt$//#if TODO</ID>
+    <ID>PreprocessingImportOrdering:InquisitorSoundConfig.kt$//#if TODO</ID>
     <ID>PreprocessingImportOrdering:LaneSwitchSoundSettings.kt$//#if TODO</ID>
     <ID>PreprocessingImportOrdering:LowHealthAlertSoundConfig.kt$//#if TODO</ID>
     <ID>PreprocessingImportOrdering:RepositoryConfig.kt$//#if TODO</ID>
